fix the behavior of updating & deleting for a specific user

+++ src/controllers/deletePersonalInfo.controller.ts
@@ -4,20 +4,26 @@ import { PersonalInfo, PrismaClient, Prisma } from "@prisma/client";
 const prisma = new PrismaClient();
 
 export const deletPersonInfo = async (req: Request, res: Response) => {
-	//NOTE: I used user_id to delete the data belongs to it which is indexed in the database
 	if (req.user) {
 		const user_id = req.user["user_id"];
+		const id = +req.params.id;
 		try {
-			const USER_INFO: PersonalInfo = await prisma.personalInfo.delete({
-				where: {
-					user_id: user_id,
-				},
-				include: {
-					urls: true,
-				},
+			const INFO_FOUND = await prisma.personalInfo.findFirstOrThrow({
+				where: { id },
 			});
-			if (USER_INFO) {
-				return res.status(200).json(USER_INFO);
+
+			if (INFO_FOUND?.user_id === user_id && INFO_FOUND?.id === id) {
+				const USER_INFO: PersonalInfo = await prisma.personalInfo.delete({
+					where: {
+						id,
+					},
+					include: {
+						urls: true,
+					},
+				});
+				if (USER_INFO) {
+					return res.status(200).json(USER_INFO);
+				}
 			}
 		} catch (error) {
 			if (error instanceof Prisma.PrismaClientKnownRequestError) {
@@ -25,7 +31,7 @@ export const deletPersonInfo = async (req: Request, res: Response) => {
 					return res.status(400).json(error.meta);
 				}
 			}
-			return res.status(404).json({ msg: "no data belongs to this user" });
+			return res.status(404).json({ msg: `no data for id: ${id}` });
 		}
 	}
 };
